# backend-main/app/openapi.yml
openapi: '3.0.2'
info:
  title: SAM API spec
  version: '1.0'
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []  
paths:
...
  /login:
    post:
      tags:
        - Authentication
      summary: Login
      description: Login with email and password
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "example@gmail.com"
                  format: email
                password:
                  type: string
                  format: password
                  example: "password"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        400:
          description: Bad Request
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/badRequest'
        401:
          description: Unauthorized
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/unauthorized'
  /check:
    get:
      tags:
        - "Authentication"
      summary: Check if the jwt provided is valid
      description: Check if the jwt provided is valid
      responses:
        200:
          description: the jwt is valid
          content:
            application/json:
              schema:
                type: string
                example: "ok"
        401:
          description: Unauthorized
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/unauthorized'

# USER
  /user/me:
    get:
      tags:
        - User
      summary: Get user information
      description: Get user information
      responses:
        200:
          description: user succesfully find
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user-public'
        401:
          description: Unauthorized
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/unauthorized'
        500:
          description: Internal Server Error
          content:
            plain/text:
              schema:
                type: string
                example: "Internal Server Error"
  /user:
    post:
      tags:
        - User
      summary: Create a new user
      description: Create a new user
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "test@gmail.com"
                password:
                  type: string
                  format: password
                  example: "password"
                name:
                  type: string
                  example: "John Doe"
                company:
                  type: string
                  example: "Company name"
                company_type:
                  type: string
                  example: "Startup"
                phone:
                  type: string
                  example: "+33612345678"
              required:
                - email
                - password
                - name
      responses:
        200:
          description: user succesfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user-public'
        400:
          description: Bad Request
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/badRequest'
        500:
          description: Internal Server Error
          content:
            plain/text:
              schema:
                type: string
                example: "Internal Server Error"
    patch:
      tags:
        - User
      summary: Update user information
      description: Update user information
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "John Doe"
                company:
                  type: string
                  example: "Company name"
                company_type:
                  type: string
                  example: "Startup"
                phone:
                  type: string
                  example: "+33612345678"
                password:
                  type: string
                  format: password
                  example: "password"
      responses:
        200:
          description: user succesfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user-public'
        400:
          description: Bad Request
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/badRequest'
        500:
          description: Internal Server Error
          content:
            plain/text:
              schema:
                type: string
                example: "Internal Server Error"
    delete:
      tags:
        - User
      summary: Delete user
      description: Delete user
      responses:
        200:
          description: user succesfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user-public'
        401:
          description: Unauthorized
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/unauthorized'
        500:
          description: Internal Server Error
          content:
            plain/text:
              schema:
                type: string
                example: "Internal Server Error"
  /user/email:
    patch:
      tags:
        - User
      summary: Update user email
      description: Update user email
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "newemail@gmail.com"
                  format: email
              required:
                - email
      responses:
        200:
          description: user succesfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user-public'
        401:
          description: Unauthorized
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/unauthorized'
        500:
          description: Internal Server Error
          content:
            plain/text:
              schema:
                type: string
                example: "Internal Server Error"
  /trouble/forgot_email:
    get:
      tags:
        - User
      summary: Send an email to the user with a link to reset his password
      description: Send an email to the user with a link to reset his password
      parameters:
        - in: query
          name: email
          schema:
            type: string
            format: email
            example: "test@gmail.com"
          required: true
          description: The email of the user
      responses:
        200:
          description: Email sended to user email
          content:
            application/json:
              schema:
                type: string
                example: Email sended to user email
        400:
          description: Bad Request
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/badRequest'
        500:
          description: Internal Server Error
          content:
            plain/text:
              schema:
                type: string
                example: "Internal Server Error"
    post:
      tags:
        - User
      summary: Reset user password
      description: Reset user password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  example: "3ab49c23-120b-4294-89df-27ad88deaf15"
                password:
                  type: string
                  format: password
                  example: "password"
              required:
                - code
                - password
      responses:
        200:
          description: user password succesfully reseted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user-public'
        400:
          description: Missing required key
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/badRequest'
        500:
          description: Internal Server Error
          content:
            plain/text:
              schema:
                type: string
                example: "Internal Server Error"
  /user/{user_id}:
    delete:
      tags:
        - User
      summary: Admin Delete user
      description: Admin Delete user
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
            example: "3ab49c23-120b-4294-89df-27ad88deaf15"
          required: true
          description: The id of the user
      responses:
        200:
          description: user succesfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user-public'
        401:
          description: Unauthorized
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/unauthorized'
        500:
          description: Internal Server Error
          content:
            plain/text:
              schema:
                type: string
                example: "Internal Server Error"

# VERIFICATION
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT 
  schemas:
...
    user-public:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "3ab49c23-120b-4294-89df-27ad88deaf15"
        email:
          type: string
          format: email
          example: "test@gmail.com"
        name:
          type: string
          example: "John Doe"
        company:
          type: string
          example: "Company name"
        credit:
          type: number
          example: 10
        company_type:
          type: string
          example: "Startup"
        phone:
          type: string
          example: "+33612345678"
        address:
          type: string
          example: "1 rue de la paix"
        email_verified:
          type: boolean
          example: true
    badRequest:
      type: string
      example: "Bad Request"
    unauthorized:
      type: string
      example: "Unauthorized, please provide JWT token"
...